feat(jest): migrate helpers to TS and expose them from the package
