+++ packages/unit-jest/src/helpers/layout-injections.ts
@@ -1,0 +1,14 @@
+/**
+ * Injections for Components with a QPage root Element
+ */
+export function qLayoutInjections() {
+  return {
+    pageContainer: true,
+    layout: {
+      header: {},
+      right: {},
+      footer: {},
+      left: {},
+    },
+  };
+}
