+++ packages/unit-jest/src/helpers/ssr-context-mock.ts
@@ -1,0 +1,11 @@
+export function ssrContextMock() {
+  return {
+    req: {
+      headers: {},
+    },
+    res: {
+      setHeader: () => undefined,
+    },
+    url: "",
+  };
+}
