+++ packages/unit-jest/src/no-typescript/jest.config.js
@@ -1,4 +1,4 @@
-const esModules = ['quasar/lang', 'lodash-es'].join('|');
+const esModules = ["quasar/lang", "lodash-es"].join("|");
 
 module.exports = {
   globals: {
@@ -39,7 +39,7 @@ module.exports = {
     "^quasar$": "<rootDir>/node_modules/quasar/dist/quasar.common.js",
     "^~/(.*)$": "<rootDir>/$1",
     "^src/(.*)$": "<rootDir>/src/$1",
-    ".*css$": "<rootDir>/test/jest/utils/stub.css",
+    ".*css$": "@quasar/quasar-app-extension-testing-unit-jest/stub.css",
   },
   transform: {
     ".*\\.vue$": "vue-jest",
